Add new Research Lab and Broadcast Tower
Also refactored the Anarchy Period to be affected by the number of
players in a civ, with a max cap defined in the config. Also rework the
bonus from the Notre Dame wonder, and add a similar effect to the
Broadcast Tower.
Finally, Reworked the entire Tech tree, scaling it into each 'Era'
rather than 'Tiers.'

+++ civcraft/src/com/avrgaming/civcraft/main/CivGlobal.java
@@ -1142,7 +1142,7 @@ public class CivGlobal {
 		return false;		
 	}
 
-	public static boolean hasTimeElapsed(SessionEntry se, int seconds) {
+	public static boolean hasTimeElapsed(SessionEntry se, double seconds) {
 		long now = System.currentTimeMillis();
 		int secondsBetween = getSecondsBetween(se.time, now);
